Added missing includes. Added warning about use of SEPARATOR in basename(...) fun at Win32
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@271 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/game/fheroes2.cpp
@@ -21,6 +21,7 @@
 #include <iostream>
 #include <string>
 #include <unistd.h>
+#include <limits.h>
 
 #include "gamedefs.h"
 #include "cmdline.h"
@@ -42,7 +43,8 @@ char *dirname(const char *path)
 {
     static char buff[PATH_MAX];
     strncpy(buff, path, PATH_MAX);
-    char *c = strrchr(buff, SEPARATOR);
+    // FIXME it won't (probably) work on Win32
+    char *c = strrchr(buff, (int)SEPARATOR);
     if(!c) strcpy(buff, ".");
     else *c = 0;
     return buff;
@@ -51,7 +53,8 @@ char *dirname(const char *path)
 char *basename(const char *path)
 {
     static char buff[PATH_MAX];
-    char *c = strrchr(path, SEPARATOR);
+    // FIXME it won't (probably) work on Win32
+    char *c = strrchr(path, (int)SEPARATOR);
     if(!c) strncpy(buff, path, PATH_MAX);
     else strcpy(buff, c);
     return buff;
